Split _rows.py into _raw_metrics.py and _stale_pr.py
- _rows.py keeps Row/Summary/Band/band_color (shared display types)
- _raw_metrics.py gets RawMetrics (compute pipe type)
- _stale_pr.py gets StalePr (stale PR detection type)
- All imports updated
