+++ git_dev_metrics/metrics/_rows.py
@@ -16,19 +16,6 @@ def band_color(band: Band) -> str:
 
 
 @dataclass(frozen=True)
-class RawMetrics:
-    cycle_time: float
-    pickup_time: float
-    review_time: float
-    pr_size: float
-    avg_lines_per_pr: float
-    pr_count: int
-    prs_per_week: float
-    reviews_given: int
-    ai_percentage: float
-
-
-@dataclass(frozen=True)
 class Row:
     name: str
     pr_count: int
@@ -42,19 +29,6 @@ class Row:
     ai_percentage: float
     health: int
     band: Band
-
-
-@dataclass(frozen=True)
-class StalePr:
-    number: int
-    title: str
-    author: str | None
-    repo: str
-    age_hours: float
-    age_days: float
-    is_draft: bool
-    is_approved: bool
-    url: str
 
 
 @dataclass(frozen=True)
